Changes were made to the 1st paragraph of the TOR: - Renamed some packages; - All stages of this paragraph of the TOR have been completed; - Added config file for MySQL deployment in docker container.
