+++ searchengine/docker-compose.yml
@@ -1,0 +1,11 @@
+version: '3'
+
+services:
+  mysql:
+    image: mysql:latest
+    container_name: mysql
+    environment:
+      - MYSQL_ROOT_PASSWORD=root
+      - MYSQL_DATABASE=search_engine
+    ports:
+      - '8081:3306'
