+++ searchengine/src/main/java/searchengine/controller/ApiController.java
@@ -1,11 +1,11 @@
-package searchengine.controllers;
+package searchengine.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import searchengine.dto.statistics.StatisticsResponse;
-import searchengine.services.StatisticsService;
+import searchengine.service.StatisticsService;
 
 @RestController
 @RequestMapping("/api")
