+++ searchengine/src/main/java/searchengine/controller/DefaultController.java
@@ -1,4 +1,4 @@
-package searchengine.controllers;
+package searchengine.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
