+++ searchengine/src/main/java/searchengine/service/StatisticsService.java
@@ -1,4 +1,4 @@
-package searchengine.services;
+package searchengine.service;
 
 import searchengine.dto.statistics.StatisticsResponse;
